% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_global.R
\name{add_global}
\alias{add_global}
\title{Adds the global p-value for a categorical variables}
\usage{
add_global(x, ...)
}
\arguments{
\item{x}{\code{fmt_regression} or \code{fmt_uni_regression} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function uses \code{\link[car]{Anova}} from the \code{car} package with \code{type = "III"} to calculate global p-values.
}
\seealso{
\code{\link{add_global.fmt_regression}}, \code{\link{add_global.fmt_uni_regression}}
}
