% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fmt_uni_regression}
\alias{print.fmt_uni_regression}
\title{Print \code{fmt_uni_regression} objects}
\usage{
\method{print}{fmt_uni_regression}(x, ...)
}
\arguments{
\item{x}{object of class \code{fmt_uni_regression} object from \code{fmt_uni_regression()} function}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print \code{fmt_uni_regression} objects
}
\examples{
fmt_uni_regression(
  trial,
  method = "glm",
  y = "response",
  method.args = list(family = binomial),
  exponentiate = TRUE
) \%>\%
  print()
}
