% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clintable-package.R
\docType{package}
\name{gtsummary-package}
\alias{gtsummary}
\alias{gtsummary-package}
\title{gtsummary: Great, Presentation-ready Data Summary Tables}
\description{
The {gtsummary} package creates
    presentation-ready tables summarizing data sets, regression models,
    and more. The code to create the tables is concise and highly
    customizable. Data frames can be summarized with many functions.
    Regression models are summarized including the reference row for
    categorical independent variables.  headers.License: MIT + file
    LICENSE
}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danieldsjoberg@gmail.com}

}
\keyword{internal}
